[LEMS-3038/improve-diner-docs] Apply documentation improvements to editors
