+++ __docs__/introduction.mdx
@@ -1,6 +1,9 @@
 import {Markdown, Meta} from "@storybook/addon-docs/blocks";
 
-<Meta title="Introduction" />
+<Meta
+    title="Introduction"
+    parameters={{docs: {toc: {headingSelector: "h2, h3"}}}}
+/>
 
 # Perseus
